requirements.txt
PyGithub
requests
Markdown
scikit-learn